+++ CHANGELOG.md
@@ -8,12 +8,12 @@
 - Buffering for in `sll_file_read` and `sll_file_read_char`
 - CLI option for enabling sandbox flags (`-S`) ([#97])
 - Implemented `sll_platform_get_default_stream_descriptor`, `sll_file_write_format`, `sll_file_write_string`, `sll_string_format`, `sll_string_format_list` and `sll_string_select`
-- String copy API: `sll_copy_string`, `sll_copy_string_null` and `SLL_COPY_STRING_NULL` ([#79])
 - Integer can now have digits split by underscores (`_`)
 - Macro constants: `SLL_PLATFORM_STREAM_INPUT`, `SLL_PLATFORM_STREAM_OUTPUT` and `SLL_PLATFORM_STREAM_ERROR`
 - Sandbox flags (`SLL_SANDBOX_FLAG_DISABLE_FILE_IO`, `SLL_SANDBOX_FLAG_ENABLE_STDIN_IO`, `SLL_SANDBOX_FLAG_ENABLE_STDOUT_IO` and `SLL_SANDBOX_FLAG_DISABLE_PATH_API`) ([#97])
 - Secure sandbox API: `sll_get_sandbox_flag` and `sll_set_sandbox_flags` ([#97])
 - `sll_api_sys_get_executable` and the `executable` variable in `sys.sll`
+- String copy API: `sll_copy_string`, `sll_copy_string_null` and `SLL_COPY_STRING_NULL` ([#79])
 - Type-check (typeof) operator (`(:?)`) ([#74])
 
 ### Changed